Remove configuration console app
