+++ lewee.sln
@@ -74,8 +74,6 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Pizzeria.Store.Application.
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Lewee.Infrastructure.Data.Tests.Unit", "tests\Lewee.Infrastructure.Data.Tests.Unit\Lewee.Infrastructure.Data.Tests.Unit.csproj", "{4907C7CA-2AA3-4E86-BCD7-09BB7AAA4A08}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Pizzeria.Configuration", "sample\Pizzeria.Configuration\Pizzeria.Configuration.csproj", "{35D290E1-04BF-4653-B5A5-9836E818B6AF}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Pizzeria.Auth", "sample\Pizzeria.Auth\Pizzeria.Auth.csproj", "{42D27FC7-14AE-4701-B6B0-A78EBA25FE1C}"
 EndProject
 Global
@@ -424,18 +422,6 @@ Global
 		{4907C7CA-2AA3-4E86-BCD7-09BB7AAA4A08}.Release|x64.Build.0 = Release|Any CPU
 		{4907C7CA-2AA3-4E86-BCD7-09BB7AAA4A08}.Release|x86.ActiveCfg = Release|Any CPU
 		{4907C7CA-2AA3-4E86-BCD7-09BB7AAA4A08}.Release|x86.Build.0 = Release|Any CPU
-		{35D290E1-04BF-4653-B5A5-9836E818B6AF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{35D290E1-04BF-4653-B5A5-9836E818B6AF}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{35D290E1-04BF-4653-B5A5-9836E818B6AF}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{35D290E1-04BF-4653-B5A5-9836E818B6AF}.Debug|x64.Build.0 = Debug|Any CPU
-		{35D290E1-04BF-4653-B5A5-9836E818B6AF}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{35D290E1-04BF-4653-B5A5-9836E818B6AF}.Debug|x86.Build.0 = Debug|Any CPU
-		{35D290E1-04BF-4653-B5A5-9836E818B6AF}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{35D290E1-04BF-4653-B5A5-9836E818B6AF}.Release|Any CPU.Build.0 = Release|Any CPU
-		{35D290E1-04BF-4653-B5A5-9836E818B6AF}.Release|x64.ActiveCfg = Release|Any CPU
-		{35D290E1-04BF-4653-B5A5-9836E818B6AF}.Release|x64.Build.0 = Release|Any CPU
-		{35D290E1-04BF-4653-B5A5-9836E818B6AF}.Release|x86.ActiveCfg = Release|Any CPU
-		{35D290E1-04BF-4653-B5A5-9836E818B6AF}.Release|x86.Build.0 = Release|Any CPU
 		{42D27FC7-14AE-4701-B6B0-A78EBA25FE1C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{42D27FC7-14AE-4701-B6B0-A78EBA25FE1C}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{42D27FC7-14AE-4701-B6B0-A78EBA25FE1C}.Debug|x64.ActiveCfg = Debug|Any CPU
