+++ sample/Pizzeria.AppHost/Program.cs
@@ -34,17 +34,11 @@ var databaseServer = Environments.IsIntegrationTesting
 var pizzaStoreDatabaseName = ServiceNames.GetPizzaStoreDatabaseName();
 var pizzaStoreDatabase = databaseServer.AddDatabase(pizzaStoreDatabaseName);
 
-var configuration = builder.AddProject<Projects.Pizzeria_Configuration>(ServiceNames.PizzeriaConfiguration)
-    .WithReference(authServer)
-    .WithReference(pizzaStoreDatabase)
-    .WaitFor(authServer)
-    .WaitFor(pizzaStoreDatabase);
-
 builder.AddProject<Projects.Pizzeria_Store_Api>(ServiceNames.PizzaStoreApi)
     .WithReference(pizzaStoreDatabase)
     .WithReference(authServer)
-    .WithEnvironment(Environments.AuthenticationSchemesBearerValidAudience, ServiceNames.PizzaStoreApi)
-    .WaitFor(configuration)
+    .WaitFor(authServer)
+    .WaitFor(pizzaStoreDatabase)
     .WithHttpHealthCheck("/health");
 
 /*
