--- sample/Pizzeria.Configuration/Services/IAuthServerConfiguration.cs
@@ -1,6 +1,0 @@
-namespace Pizzeria.Configuration.Services;
-
-public interface IAuthServerConfiguration
-{
-    Task ConfigureAsync();
-}
