--- sample/Pizzeria.Configuration/Services/IDatabaseConfigurationService.cs
@@ -1,8 +1,0 @@
-namespace Pizzeria.Configuration.Services;
-
-public interface IDatabaseConfigurationService
-{
-    Task MigrateAsync();
-
-    Task SeedDataAsync();
-}
